fix(app): simplified check licence
