+++ app/services/return-requirements/check-licence-ended.service.js
@@ -1,18 +1,50 @@
-// check-licence-ended.service.js
-
 'use strict'
 
 /**
- * Service to check if a licence has ended.
- * This service checks the 'ends' date of the licence data and determines if the licence has ended.
+ * Fetches and checks data needed for the view `/return-requirements/{sessionId}/check-your-answers` page
+ * @module CheckLicenceService
+ */
+
+const LicenceModel = require('../../models/licence.model.js')
+
+/**
+ * Fetch the matching licence and return data needed for the check your answers page
+ *
+ * Was built to provide the data needed for the '/return-requirements/{id}/check-your-answers' page
+ *
+ * @param {string} id The UUID for the licence to fetch
+ *
+ * @returns {Promise<Object>} the data needed for "check your answers" page when user submits for approval
+ */
+async function go (id) {
+  const licence = await _fetchLicence(id)
+  const licenceEnded = _checkLicenceEnded(licence)
+  return licenceEnded
+}
+
+async function _fetchLicence (id) {
+  return LicenceModel.query()
+    .findById(id)
+    .select([
+      'expiredDate',
+      'id',
+      'lapsedDate',
+      'licenceRef',
+      'revokedDate',
+      'startDate'
+    ])
+}
+
+/**
+ * Check if a licence has ended.
  * If it's not within range, we assume it's either lapsed, expired or revoked
  * It returns true if the licence has ended, false otherwise.
  *
  * @param {object} licenceData - The licence data object
  * @returns {boolean} - True if the licence has ended, false otherwise
  */
-async function checkLicenceEnded (licenceData) {
-  const { ends } = licenceData
+function _checkLicenceEnded (licence) {
+  const { $ends: ends } = licence
 
   if (!ends) {
     return false
@@ -26,5 +58,5 @@ async function checkLicenceEnded (licenceData) {
 }
 
 module.exports = {
-  checkLicenceEnded
+  go
 }
