+++ test/services/return-requirements/check-licence-ended.service.test.js
@@ -4,51 +4,30 @@
 const Lab = require('@hapi/lab')
 const Code = require('@hapi/code')
 
-const { describe, it } = exports.lab = Lab.script()
+const { describe, it, beforeEach } = exports.lab = Lab.script()
 const { expect } = Code
+
+// Test helpers
+const DatabaseSupport = require('../../support/database.js')
+const LicenceHelper = require('../../support/helpers/licence.helper.js')
 
 // Thing under test
 const CheckLicenceEndedService = require('../../../app/services/return-requirements/check-licence-ended.service.js')
 
-describe('Check Licence Ended Service', () => {
-  describe('when the licence has not ended', () => {
-    it('returns false', async () => {
-      // tomorrow
-      const futureDate = new Date()
-      futureDate.setDate(futureDate.getDate() + 1)
+describe('CheckLicenceEndedService', () => {
+  let licence
 
-      const licenceData = {
-        ends: {
-          date: futureDate.toISOString()
-        }
-      }
-
-      const result = await CheckLicenceEndedService.checkLicenceEnded(licenceData)
-      expect(result).to.be.false()
-    })
+  beforeEach(async () => {
+    await DatabaseSupport.clean()
   })
 
-  describe('when the licence has ended', () => {
-    it('returns true', async () => {
-      // a date in the past
-      const pastDate = new Date('2024-02-14')
-
-      const licenceData = {
-        ends: {
-          date: pastDate.toISOString()
-        }
-      }
-
-      const result = await CheckLicenceEndedService.checkLicenceEnded(licenceData)
-      expect(result).to.be.true()
+  describe('go', () => {
+    beforeEach(async () => {
+      licence = await LicenceHelper.add()
     })
-  })
 
-  describe('when the licence ends data is not available', () => {
-    it('returns false', async () => {
-      const licenceData = {}
-
-      const result = await CheckLicenceEndedService.checkLicenceEnded(licenceData)
+    it('fetches licence data correctly', async () => {
+      const result = await CheckLicenceEndedService.go(licence.id)
       expect(result).to.be.false()
     })
   })
